Fix exec::Strings  allocation pattern
the Strings collection is used for variable length payload data in
aggregates. It would previously allocate the exact size needed to
cover the string and the next pointer and the 8 byte gap at the head
of a continued HashStrigAllocator allocation. Each append to Strings
would make a single string. These are typically small, e.g. 15-20b of
string and 16b of overhead.

This falls in the smallest allocation class of HashStringAllocator,
under 72b. Making many such small allocations causes ends of blocks to
be added to the smallest free list. The ends are 17-27b and are always
too small for the smallest possible allocation from Strings, which is
13 + 16 = 29b. The allocations from Strings also always fall in the
smallest class but loop through 30+K of ends of blocks that are too
small. This makes all the time go to scanning the free list.

Now therefore, one fix is to make Strings allocate larger blocks,
e.g. 4 * max string size, which was the original intent, but never did
work. This causes ends of blocks still to be created from time to time
but they go to the small list which Strings does not use, except for
the first two allocations. Strings is also thereby slightly more
efficient since it does not incur the 16b overhead for every single
string.

We round up sizes to the minimum size of HashStringAllocator free
lists. This means no scanning through free lists since all blocks will
be large enough.

We track the largest size available in each free list of
HashStringAllocator. This is updated when an entry is added. This is
not updated when removing an entry. After traversing the whole free
list and failing to find a fit, we know the largest size and update
the record.

+++ velox/exec/Strings.h
@@ -32,7 +32,7 @@ struct Strings {
 
   /// Maximum size of a string passed to 'append'. Used to calculate the amount
   /// of space to reserve for future strings.
-  size_t maxStringSize = 0;
+  int32_t maxStringSize = 0;
 
   /// Copies the string into contiguous memory allocated via
   /// HashStringAllocator. Returns StringView over the copy.
